+++ src/Transaction/InitialTransaction.php
@@ -32,7 +32,7 @@
 
 namespace Wirecard\PaymentSdk\Transaction;
 
-use Wirecard\PaymentSdk\Money;
+use Wirecard\PaymentSdk\Entity\Money;
 
 /**
  * Class Transaction
