+++ test/Entity/MoneyUTest.php
@@ -30,9 +30,9 @@
  * Please do not use the plugin if you do not agree to these terms of use!
  */
 
-namespace WirecardTest\PaymentSdk;
+namespace WirecardTest\PaymentSdk\Entity;
 
-use Wirecard\PaymentSdk\Money;
+use Wirecard\PaymentSdk\Entity\Money;
 
 class MoneyUTest extends \PHPUnit_Framework_TestCase
 {
