+++ test/Transaction/CreditCardTransactionUTest.php
@@ -33,7 +33,7 @@
 namespace WirecardTest\PaymentSdk\Transaction;
 
 use Wirecard\PaymentSdk\Transaction\CreditCardTransaction;
-use Wirecard\PaymentSdk\Money;
+use Wirecard\PaymentSdk\Entity\Money;
 
 class CreditCardTransactionUTest extends \PHPUnit_Framework_TestCase
 {
